Refactor Tracker for external command control
The `Tracker` and `Ldrs` classes have been refactored to remove the internal implementation of autonomous LDR-based tracking logic. Previously, `Tracker` evaluated LDR differences and initiated `deploy` or `retract` actions based on its internal `_isAutoMode` state. The `Ldrs` class also contained comparison methods (e.g., `isDayUpBrighterThanDayDown`).

This change removes `_isAutoMode`, `isAutoMode()`, `setAutoMode()`, `scan()`, and the LDR comparison logic from `Tracker::updateAutoMode` (now `Tracker::update`). The LDR comparison methods have also been removed from `Ldrs`.

The `Tracker` now acts as a command-driven component, solely providing `deploy`, `retract`, and `stop` actions. The responsibility for deciding *when* to track is moved to an external entity, promoting a clearer separation of concerns.

Additionally, various `int` type parameters and member variables related to ADC/PWM resolutions, percentages, and motor speeds have been refined to `uint16_t` or `uint8_t` for better type safety and memory efficiency.

+++ src/driver/motor.cpp
@@ -4,7 +4,7 @@
 #include "setting.h"
 #include <ArduinoLog.h>
 
-Motor::Motor(uint8_t in1, uint8_t in2, uint8_t en, int pwmResolution)
+Motor::Motor(uint8_t in1, uint8_t in2, uint8_t en, uint8_t pwmResolution)
     : _in1(in1), _in2(in2), _en(en), _pwmResolution(pwmResolution) {}
 
 void Motor::init() {
